made datagrid a component
