+++ src/pages/midwife/eligibleCouples/viewAll.jsx
@@ -15,29 +15,8 @@ import { errorType, Toast } from "@/components/toast";
 import Heading from "@/components/ui/heading";
 import { Link, useNavigate } from "react-router-dom";
 import { useTitle } from "@/hooks/useTitle";
-import Popup from "reactjs-popup";
-
-const StripedDataGrid = styled(DataGrid)(({ theme }) => {
-  return {
-    [`& .${gridClasses.row}.even`]: {
-      backgroundColor: theme.palette.mode === "dark" ? "#333333" : "#FAEDFF",
-    },
-    [`& .${gridClasses.row}.odd`]: {
-      backgroundColor: theme.palette.mode === "dark" ? "#444444" : "#ffffff",
-    },
-    border: "none",
-
-    "& .MuiDataGrid-cell:focus": {
-      outline: "none",
-    },
-    "& .MuiDataGrid-cell:focus-within": {
-      outline: "none",
-    },
-    "& .MuiDataGrid-columnHeaders": {
-      backgroundColor: theme.palette.mode === "dark" ? "#000000" : "#555555", // Light vs dark mode header color
-    },
-  };
-});
+import EligiblePopup from "@/components/eligiblePopup";
+import { StripedDataGrid } from "@/components/StripedDataGrid";
 
 function QuickSearchToolbar() {
   return (
@@ -45,9 +24,9 @@ function QuickSearchToolbar() {
       sx={{
         p: 0.5,
         pb: 0,
-        m: 2,
+        mb: 2,
         display: "flex",
-        justifyContent: "flex-end",
+        justifyContent: "flex-start",
       }}
     >
       <GridToolbarQuickFilter
@@ -68,7 +47,7 @@ function QuickSearchToolbar() {
 }
 
 const columns = [
-  { field: "id", headerName: "E. Couple ID", width: 70 },
+  { field: "id", headerName: "E. Couple ID", width: 100,},
   {
     field: "name",
     headerName: "Mother / Father",
@@ -107,7 +86,7 @@ const columns = [
   {
     field: "status",
     headerName: "Status",
-    flex: 1,
+    width: 170,
     renderCell: (params) => {
       const isEligible = params.value.role === "ELIGIBLE";
       return (
@@ -146,34 +125,8 @@ const columns = [
 const eligibleCouples = () => {
   useTitle("Eligible Couples");
   const [rows, setRows] = useState([]);
-  const [errors, setErrors] = useState({});
-  const [showExisting, setShowExisting] = useState(false);
-  const [showNew, setShowNew] = useState(false);
   const navigate = useNavigate();
   const { t } = useTranslation("eligibleCouples");
-
-  const handleInputChange = (e) => {
-    const { name, value } = e.target;
-    // setFormData({ ...formData, [name]: value });
-    setErrors({ ...errors, [name]: validateField(name, value) });
-  };
-
-  const validateField = (name, value) => {
-    switch (name) {
-      case "firstName":
-        if (!value) return "First name is required";
-        break;
-      case "lastName":
-        if (!value) return "Last name is required";
-        break;
-      case "email":
-        if (!value) return "Email address is required";
-        break;
-      default:
-        break;
-    }
-    return "";
-  };
 
   useEffect(() => {
     const fetchEligibleListForMidwife = async () => {
@@ -213,91 +166,8 @@ const eligibleCouples = () => {
     <div className="content-container">
       <Heading title={t("title")} />
 
-      <div className="flex flex-col items-end mt-10">
-        <Popup
-          trigger={
-            <Button className="bg-[#6F0096] h-10 flexbox items-center ">
-              {t("add")}
-              <HiOutlinePlusSm className="ml-2 h-5 w-5" />
-            </Button>
-          }
-          modal
-          nested
-          // overlayClassName="overlay"
-          overlayStyle={{
-            backgroundColor: "rgba(0, 0, 0, 0.6)",
-          }}
-          // className="centered-popup"
-        >
-          <div className="bg-white dark:bg-dark-popup w-fit p-10 rounded-md">
-            <p className="m-10 mt-0 text-xl font-semibold">
-              Existing User or new couple?
-            </p>
-            {!showExisting && !showNew && (
-              <div className="buttons flex gap-12 w-full justify-center">
-                <Button className="" onClick={() => setShowExisting(true)}>
-                  Existing
-                </Button>
-                <Button onClick={() => setShowNew(true)}>New</Button>
-              </div>
-            )}
-            {showExisting && (
-              <div className="existing flex flex-col items-center gap-8">
-                <TextField
-                  label="Email Address"
-                  type="email"
-                  required
-                  variant="standard"
-                  fullWidth
-                  name="email"
-                  onChange={handleInputChange}
-                  error={errors.email ? true : false}
-                  helperText={errors.email ? errors.email : ""}
-                  // className="rounded"
-                ></TextField>
-
-                <Button className="px-10">Submit</Button>
-              </div>
-            )}
-            {showNew && (
-              <div className="new flex flex-col items-center gap-8">
-                <TextField
-                  label="First Name"
-                  required
-                  variant="standard"
-                  fullWidth
-                  name="firstName"
-                  onChange={handleInputChange}
-                  error={errors.firstName ? true : false}
-                  helperText={errors.firstName ? errors.firstName : ""}
-                ></TextField>
-                <TextField
-                  label="Last Name"
-                  required
-                  variant="standard"
-                  fullWidth
-                  name="lastName"
-                  onChange={handleInputChange}
-                  error={errors.lastName ? true : false}
-                  helperText={errors.lastName ? errors.lastName : ""}
-                ></TextField>
-                <TextField
-                  label="Email Address"
-                  type="email"
-                  required
-                  variant="standard"
-                  fullWidth
-                  name="email"
-                  onChange={handleInputChange}
-                  error={errors.email ? true : false}
-                  helperText={errors.email ? errors.email : ""}
-                ></TextField>
-
-                <Button className="px-10">Submit</Button>
-              </div>
-            )}
-          </div>
-        </Popup>
+      <div className="flex flex-col items-end">
+        <EligiblePopup addButton={t("add")}></EligiblePopup>
 
         {/* clinics table */}
         <div style={{ height: "100%", width: "100%" }}>
